refactor: group UserDetailPage modals and table internals
