+++ src/routes/users.tsx
@@ -1,6 +1,6 @@
 import BackToHomeLink from "@/components/display/GameServer/GameServerDetailPageLayout/BackToHomeLink";
 import { GameServerNotFoundPage } from "@/components/display/GameServer/GameServerNotFoundPage/GameServerNotFoundPage";
-import UserTable from "@/components/display/UserManagement/UserDetailPage/UserTable";
+import UserTable from "@/components/display/UserManagement/UserDetailPage/UserTable/UserTable";
 import { AuthContext } from "@/components/technical/Providers/AuthProvider/AuthProvider";
 import { ThemeContext, ThemeOptions } from "@/components/technical/Providers/ThemeProvider.tsx";
 import { createFileRoute } from "@tanstack/react-router";
